Cambio de tipo de variable

+++ AlazarSoft/src/co/edu/unbosque/controller/Controller.java
@@ -29,7 +29,7 @@ public class Controller implements ActionListener {
 
 		if (entradas[0].equals("0")) {
 			casaApuestas.guardarConfiguracionCasaDeApuestas(entradas[1], Integer.parseInt(entradas[2]),
-					Double.parseDouble(entradas[3]));
+					Long.parseLong(entradas[3]));
 			vista.mostrarMensajeInformacion(entradas[4]);
 		} else {
 			vista.mostrarMensajeError(entradas[1]);
